+++ ls_session.h
@@ -19,7 +19,7 @@ typedef struct ls_session_s {
     // const map<string, void*> *settings;// 引用master的 TODO ？？ 可有可无
 
     const ls_task_script_t* script;// TODO script只读
-    int script_cur;
+    size_t script_cur;
 
     map<string/* plugin_name */, void*> states;
     ls_task_var_t cur_vars;// const map<string, ls_var_t> vars;// session需要根据此初始化变量cur_vars、每次执行API前获取需要的变量更新cur_vars
